requirements.txt
acme==1.18.0
asgiref==3.3.4
asymmetric-jwt-auth==0.5.2
Babel==2.9.1
boto3==1.18.28
botocore==1.21.28
certbot==1.18.0
certbot-django==0.2.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
ConfigArgParse==1.5.2
configobj==5.0.6
cryptography==3.4.7
defusedxml==0.7.1
distro==1.6.0
dj-database-url==0.5.0
Django==3.2.4
django-allauth==0.44.0
django-bmemcached==0.3.0
django-ckeditor==6.1.0
django-cors-headers==3.8.0
django-countries==7.2.1
django-crispy-forms==1.12.0
django-db-log==2.2.1
django-db-logger==0.1.10
django-debug-toolbar==3.2.1
django-environ==0.4.5
django-heroku==0.3.1
django-js-asset==1.2.2
django-mptt==0.13.2
django-phonenumber-field==5.2.0
django-s3-storage==0.13.4
django-stdimage==5.3.0
django-storages==1.11.1
djangorestframework==3.12.4
gunicorn==20.1.0
idna==2.10
jmespath==0.10.0
josepy==1.8.0
oauthlib==3.1.1
parsedatetime==2.6
phonenumbers==8.12.30
Pillow==8.2.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
pyOpenSSL==20.0.1
pyRFC3339==1.1
python-binary-memcached==0.30.1
python-dateutil==2.8.2
python-memcached==1.59
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
s3transfer==0.5.0
sentry-sdk==1.3.1
six==1.16.0
sqlparse==0.4.1
uhashring==2.1
urllib3==1.26.5
whitenoise==5.3.0
zope.component==5.0.1
zope.event==4.5.0
zope.hookable==5.1.0
zope.interface==5.4.0
